update for pecl installer

+++ pecl_install_php_extensions.sh
@@ -31,25 +31,30 @@ PHP_INI_DIR=$(php --ini | grep "Scan for additional .ini files in" | cut -d':' -
 ## pecl install imagick memcached mongodb oauth xdebug
 ## use proxy: curl -v --socks5-hostname 127.0.0.1:55880
 apt install -y libmagickwand-dev libmemcached-dev zlib1g-dev --no-install-recommends && \
-    mkdir -p /tmp/pecl_downloads && cd /tmp/pecl_downloads && \
+    mkdir -p /tmp/pecl_downloads && \
+    : && \
+    cd /tmp/pecl_downloads && \
     curl -SL http://pecl.php.net/get/imagick -o imagick.tgz && \
     curl -SL http://pecl.php.net/get/memcached -o memcached.tgz && \
     curl -SL http://pecl.php.net/get/mongodb -o mongodb.tgz && \
     curl -SL http://pecl.php.net/get/oauth -o oauth.tgz && \
     curl -SL http://pecl.php.net/get/redis -o redis.tgz && \
     curl -SL http://pecl.php.net/get/xdebug -o xdebug.tgz && \
+    : && \
     printf "\n" | pecl install --force imagick.tgz && \
     printf "\n" | pecl install --force memcached.tgz && \
     printf "\n" | pecl install --force mongodb.tgz && \
     printf "\n" | pecl install --force oauth.tgz && \
     printf "\n" | pecl install --force redis.tgz && \
     printf "\n" | pecl install --force xdebug.tgz && \
+    : && \
     echo 'extension=imagick.so' > $PHP_INI_DIR/90-imagick.ini && \
     echo 'extension=memcached.so' > $PHP_INI_DIR/90-memcached.ini && \
     echo 'extension=mongodb.so' > $PHP_INI_DIR/90-mongodb.ini && \
     echo 'extension=oauth.so' > $PHP_INI_DIR/90-oauth.ini && \
     echo 'extension=redis.so' > $PHP_INI_DIR/90-redis.ini && \
     echo 'zend_extension=xdebug.so' > $PHP_INI_DIR/90-xdebug.ini && \
+    : && \
     rm -rf /tmp/pecl_downloads
 
 ## swoole
@@ -57,11 +62,14 @@ apt install -y libmagickwand-dev libmemcached-dev zlib1g-dev --no-install-recomm
 ## hiredis( for swoole )
 ## https://github.com/redis/hiredis
 apt install -y libpq-dev nghttp2 libnghttp2-dev --no-install-recommends && \
-    mkdir -p /tmp/pecl_downloads && cd /tmp && \
+    mkdir -p /tmp/pecl_downloads && \
+    : && \
+    cd /tmp && \
     curl -o ./pecl_downloads/hiredis.tar.gz https://github.com/redis/hiredis/archive/master.tar.gz -L && \
     tar zxvf ./pecl_downloads/hiredis.tar.gz && \
     mv hiredis* hiredis && cd hiredis && \
     make -j && make install && ldconfig && \
+    : && \
     cd /tmp && \
     curl -o ./pecl_downloads/swoole.tar.gz https://github.com/swoole/swoole-src/archive/master.tar.gz -L && \
     tar zxvf ./pecl_downloads/swoole.tar.gz && \
@@ -76,18 +84,31 @@ apt install -y libpq-dev nghttp2 libnghttp2-dev --no-install-recommends && \
         --enable-coroutine-postgresql && \
     make clean && make && make install && \
     echo 'extension=swoole.so' > $PHP_INI_DIR/90-swoole.ini && \
+    : && \
     rm -rf /tmp/pecl_downloads /tmp/hiredis /tmp/swoole-src
 
 ## Phalcon
 ## https://github.com/phalcon/cphalcon
 apt install -y php${PHP_VERSION}-dev libpcre3-dev gcc make re2c --no-install-recommends && \
-    mkdir -p /tmp/pecl_downloads && cd /tmp && \
+    mkdir -p /tmp/pecl_downloads && \
+    : && \
+    cd /tmp && \
+    curl -o ./pecl_downloads/php-psr.tar.gz https://github.com/jbboehr/php-psr/archive/master.tar.gz -L && \
+    tar zxvf ./pecl_downloads/php-psr.tar.gz && \
+    mv php-psr* php-psr && cd php-psr && \
+    /usr/bin/phpize${PHP_VERSION} && \
+    ./configure --with-php-config=/usr/bin/php-config${PHP_VERSION} && \
+    make && make test && make install && \
+    echo 'extension=psr.so' > $PHP_INI_DIR/50-psr.ini && \
+    : && \
+    cd /tmp && \
     curl -o ./pecl_downloads/cphalcon.tar.gz https://github.com/phalcon/cphalcon/archive/master.tar.gz -L && \
     tar zxvf ./pecl_downloads/cphalcon.tar.gz && \
     mv cphalcon* cphalcon && cd cphalcon/build && \
     ./install --phpize /usr/bin/phpize${PHP_VERSION} --php-config /usr/bin/php-config${PHP_VERSION} && \
     echo 'extension=phalcon.so' > $PHP_INI_DIR/90-phalcon.ini && \
-    rm -rf /tmp/pecl_downloads /tmp/cphalcon
+    : && \
+    rm -rf /tmp/pecl_downloads /tmp/php-psr /tmp/cphalcon
 
 ## PDFlib
 ## https://www.pdflib.com/download/pdflib-product-family/
@@ -99,7 +120,7 @@ if [[ "$PDFlib_CURRENT_VER" != "$PDFlib_REMOTE_VER" ]]; then
         curl -o pdflib.tar.gz https://www.pdflib.com/binaries/PDFlib/912/PDFlib-9.1.2p1-Linux-x86_64-php.tar.gz -L && \
         tar -xvf pdflib.tar.gz && \
         mv PDFlib-* pdflib
-    
+
     if [[ -d "/tmp/pdflib/bind/php/php-${PDFlib_BIN_VER}-nts" ]]; then
         cp /tmp/pdflib/bind/php/php-${PDFlib_BIN_VER}-nts/php_pdflib.so $PHP_EXT_DIR && \
         echo 'extension=php_pdflib.so' > $PHP_INI_DIR/90-pdflib.ini
@@ -159,16 +180,15 @@ if [[ "$ORACLE_INSTANT_EXIST" == "no" && "$ORACLE_INSTANT_CLIENT" == "18c" ]]; t
         curl -SL -O https://github.com/epoweripione/oracle-instantclient-18/raw/master/instantclient-sdk-linux.x64-18.3.0.0.0dbru.zip && \
         curl -SL -O https://github.com/epoweripione/oracle-instantclient-18/raw/master/instantclient-sqlplus-linux.x64-18.3.0.0.0dbru.zip && \
         curl -SL -O https://github.com/epoweripione/oracle-instantclient-18/raw/master/instantclient-tools-linux.x64-18.3.0.0.0dbru.zip && \
+        : && \
         unzip instantclient-basic-linux.x64-18.3.0.0.0dbru.zip && \
         unzip instantclient-sdk-linux.x64-18.3.0.0.0dbru.zip && \
         unzip instantclient-sqlplus-linux.x64-18.3.0.0.0dbru.zip && \
         unzip instantclient-tools-linux.x64-18.3.0.0.0dbru.zip && \
-        echo /opt/oracle/instantclient_18_3 > /etc/ld.so.conf.d/oracle-instantclient18.3 && \
-        ldconfig && \
         : && \
-        export ORACLE_HOME="/opt/oracle/instantclient_18_3" && \
-        if [[ -z "$LD_LIBRARY_PATH" ]]; then export LD_LIBRARY_PATH=$ORACLE_HOME; else export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:$ORACLE_HOME; fi && \
-        export PATH=$PATH:$ORACLE_HOME && \
+        echo /opt/oracle/instantclient_18_3 > /etc/ld.so.conf.d/oracle-instantclient18.3 && \
+        : && \
+        ldconfig && \
         : && \
         rm -rf /opt/oracle/*.zip
 elif [[ "$ORACLE_INSTANT_EXIST" == "no" && "$ORACLE_INSTANT_CLIENT" == "12c" ]]; then
@@ -177,30 +197,64 @@ elif [[ "$ORACLE_INSTANT_EXIST" == "no" && "$ORACLE_INSTANT_CLIENT" == "12c" ]];
         curl -SL -O https://github.com/epoweripione/oracle-instantclient/raw/master/instantclient-sdk-linux.x64-12.2.0.1.0.zip && \
         curl -SL -O https://github.com/epoweripione/oracle-instantclient/raw/master/instantclient-sqlplus-linux.x64-12.2.0.1.0.zip && \
         curl -SL -O https://github.com/epoweripione/oracle-instantclient/raw/master/instantclient-tools-linux.x64-12.2.0.1.0.zip && \
+        : && \
         unzip instantclient-basic-linux.x64-12.2.0.1.0.zip && \
         unzip instantclient-sdk-linux.x64-12.2.0.1.0.zip && \
         unzip instantclient-sqlplus-linux.x64-12.2.0.1.0.zip && \
         unzip instantclient-tools-linux.x64-12.2.0.1.0.zip && \
+        : && \
         ln -s /opt/oracle/instantclient_12_2/libclntsh.so.12.1 /opt/oracle/instantclient_12_2/libclntsh.so && \
         ln -s /opt/oracle/instantclient_12_2/libocci.so.12.1 /opt/oracle/instantclient_12_2/libocci.so && \
         echo /opt/oracle/instantclient_12_2 > /etc/ld.so.conf.d/oracle-instantclient12.2 && \
-        ldconfig && \
         : && \
-        export ORACLE_HOME="/opt/oracle/instantclient_12_2" && \
-        if [[ -z "$LD_LIBRARY_PATH" ]]; then export LD_LIBRARY_PATH=$ORACLE_HOME; else export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:$ORACLE_HOME; fi && \
-        export PATH=$PATH:$ORACLE_HOME && \
+        ldconfig && \
         : && \
         rm -rf /opt/oracle/*.zip
 fi
 
-if [[ ! -s "$PHP_INI_DIR/90-oci8.ini" ]]; then
-    apt install -y build-essential libaio1 --no-install-recommends && \
-        mkdir -p /tmp/pecl_downloads && cd /tmp/pecl_downloads && \
-        curl -SL http://pecl.php.net/get/oci8 -o oci8.tgz && \
-        printf "instantclient,$ORACLE_HOME\n" | pecl install --force oci8.tgz && \
-        echo 'extension=oci8.so' > $PHP_INI_DIR/90-oci8.ini && \
-        : && \
-        rm -rf /tmp/pecl_downloads
+# Oracle Instant Client
+if [[ -d "/opt/oracle/instantclient_18_3" ]]; then
+    ORACLE_HOME="/opt/oracle/instantclient_18_3"
+elif [[ -d "/opt/oracle/instantclient_12_2" ]]; then
+    ORACLE_HOME="/opt/oracle/instantclient_12_2"
+fi
+
+# if [[ -d "/opt/oracle/instantclient_18_3" ]]; then
+#     export ORACLE_HOME="/opt/oracle/instantclient_18_3"
+# elif [[ -d "/opt/oracle/instantclient_12_2" ]]; then
+#     export ORACLE_HOME="/opt/oracle/instantclient_12_2"
+# fi
+
+# if [[ -n "$ORACLE_HOME" ]]; then
+#     if [[ -z "$LD_LIBRARY_PATH" ]]; then
+#         export LD_LIBRARY_PATH=$ORACLE_HOME
+#     else
+#         export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:$ORACLE_HOME
+#     fi
+
+#     export PATH=$PATH:$ORACLE_HOME
+
+#     if [[ -x "$(command -v rlwrap)" ]]; then
+#         alias sqlplus="rlwrap sqlplus"
+#         alias rman="rlwrap rman"
+#         alias asmcmd="rlwrap asmcmd"
+#         alias ggsci="rlwrap ggsci"
+#     fi
+# fi
+
+# oci8
+if ls /etc/ld.so.conf.d/oracle-instantclient* >/dev/null 2>&1; then
+    if [[ ! -s "$PHP_INI_DIR/90-oci8.ini" ]]; then
+        apt install -y build-essential libaio1 --no-install-recommends && \
+            mkdir -p /tmp/pecl_downloads && \
+            : && \
+            cd /tmp/pecl_downloads && \
+            curl -SL http://pecl.php.net/get/oci8 -o oci8.tgz && \
+            printf "instantclient,$ORACLE_HOME\n" | pecl install --force oci8.tgz && \
+            echo 'extension=oci8.so' > $PHP_INI_DIR/90-oci8.ini && \
+            : && \
+            rm -rf /tmp/pecl_downloads
+    fi
 fi
 
 ## How to check php extensions which shared libraries depends on
